modify menu display

+++ models/account.go
@@ -1,8 +1,8 @@
 package models
 
 type Account struct {
-	Id      int    `json:"id"`
-	Name    string `json:"name"`
-	Pin     string `json:"pin"`
-	Balance int    `json:"balance"`
+	Id      int     `json:"id"`
+	Name    string  `json:"name"`
+	Pin     string  `json:"pin"`
+	Balance float64 `json:"balance"`
 }
